+++ view/comments/run.go
@@ -6,7 +6,7 @@ import (
 )
 
 func Run(thread *comment.Thread, lastVisited int64, commentWidth, indent int, enableNerdFonts bool) error {
-	return pane.RunStandalone(func(width, height int) pane.View {
+	return pane.RunStandalone(thread.Title, func(width, height int) pane.View {
 		m := New(thread, lastVisited, commentWidth, indent, enableNerdFonts, width, height)
 		m.DisableAppKeys()